fft funcionando
