+++ index.html
@@ -14,7 +14,7 @@
              <script src= "./js/libs/huffman.js" type="application/javascript"></script>
              <!-- <script src= "./js/libs/fft.js" type="application/javascript"></script>
              <script src= "./js/libs/fft_image.js" type="application/javascript"></script>
-             <script src= "./js/libs/fft.js" type="application/javascript"></script> -->
+             <script src= "./js/libs/fft.js" type="application/javascript"></script> -->           
              
              <link rel="stylesheet" href="./css/libs/bootstrap.css">
